requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.4
python-multipart==0.0.22

spacy==3.8.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl

PyPDF2==3.0.1
python-docx==1.2.0
pytesseract==0.3.13
pillow==11.0.0
pdfplumber==0.11.9

transformers==4.49.0
torch==2.5.1
sentence-transformers==2.7.0
scikit-learn==1.3.2
numpy==1.26.4

ollama==0.6.1

requests==2.32.5
pandas==2.1.4